add docs and revert tests
